guru_gedara_backend_v1.0-20250707T052238Z-1-001/guru_gedara_backend_v1.0/requirements.txt
flask
flask-cors
tensorflow==2.14.0
numpy==1.26.4
scikit-learn==1.3.0
matplotlib==3.7.2
opencv-python==4.8.1.78
pillow==9.4.0
joblib==1.2.0
mediapipe==0.10.20
msvc-runtime==14.42.34433